start over
Originally, the goal was to provide feature partiy with the Celluloid
future. The structure of this implementation is difficult to test
because none of the future's dependencies can be injected at test
time; the *arguements parameter makes it difficult to tell which
paramters are config objects and which are to be passed into the
block. This led to some uncomfortable maneuvering in the specs; usage of
things like stub_const.

In order to provide a more flexible API for testing, this commit burns
this project to the ground. The objective is to take what was learned
the first go around and implement a highly testable forking future.