rename Future internals to expose Promise concept
Technically the concept of a Future is simply a bucket where some
value will be set at a point in future time. It is the Future's
Promise which is responsible for calculating its value and making that
value available to the waiting Future. This change renames the
internals to expose this fact.