isolate process forking/management logic
The fact that process forking is used to achieve concurrent processing
of futures is neither a concern of the Future (the value) nor the
Promise (logic which computes the value). As such, this extracts all
of the messy process management bits into their own class which acts
as a middle man between the logic of the Promise and the value bucket
of the Future.