move subprocess error handling
When an error occurs while executing the promise in a subprocess, the
error needs to be captured and serialized across the pipe back to the
parent process. The promise itsef doesn't necessarily care that this
special handling needs to occur; the special handling is necessary
because the Promise is evaluated in a child process. As such, it seems
more approprate for the process managing object to coordinate the
special marshalling of the error state across processes.